+++ src/subcommands/profile/configure.rs
@@ -1,7 +1,7 @@
-use crate::subcommands::profile::pick_mod_loader;
+use crate::subcommands::profile::{pick_minecraft_version, pick_mod_loader};
 use anyhow::Result;
 use dialoguer::{Input, Select};
-use libium::{config, file_picker, misc};
+use libium::{config, file_picker};
 use std::path::PathBuf;
 
 pub async fn configure(
@@ -57,22 +57,8 @@ pub async fn configure(
                             profile.output_dir = dir;
                         }
                     },
-                    1 => {
-                        // Let user pick mc version from latest 10 versions
-                        let mut versions = misc::get_major_mc_versions(10).await?;
-                        let index = Select::with_theme(&*crate::THEME)
-                            .with_prompt("Select a Minecraft version")
-                            .items(&versions)
-                            .default(0)
-                            .interact_opt()?;
-                        if let Some(i) = index {
-                            profile.game_version = versions.swap_remove(i);
-                        }
-                    },
-                    2 => {
-                        // Let user pick mod loader
-                        profile.mod_loader = pick_mod_loader(Some(&profile.mod_loader))?;
-                    },
+                    1 => profile.game_version = pick_minecraft_version().await?,
+                    2 => profile.mod_loader = pick_mod_loader(Some(&profile.mod_loader))?,
                     3 => {
                         let name = Input::with_theme(&*crate::THEME)
                             .with_prompt("Change the profile's name")